+++ src/parsing/mod.rs
@@ -1,16 +1,13 @@
 #![allow(dead_code)]
 
 pub(crate) mod address;
-pub(crate) mod base64;
 pub(crate) mod character_sets;
 pub(crate) mod combinators;
 pub(crate) mod common;
 pub(crate) mod fields;
 pub(crate) mod message;
 #[cfg(feature = "mime")]
-pub(crate) mod mime_fields;
-#[cfg(feature = "mime")]
-pub(crate) mod quoted_printables;
+pub(crate) mod mime;
 pub(crate) mod quoted_string;
 pub(crate) mod time;
 pub(crate) mod whitespaces;
